+++ src/mrmcc3/boot_soy.clj
@@ -13,10 +13,11 @@
                   com.google.code.gson/gson]]])
 
 (deftask soy
-  "replace all soy templates with a render function in the fileset
-  metadata (key ::render) for use in tasks that follow. the render
-  function takes the qualified template name as a string, a map
-  of data and returns the rendered template as a string."
+  "compile all soy templates and provide a render function
+  in the fileset metadata (key ::render) for use in tasks that follow.
+  The render function takes the qualified template name as a string,
+  a map of data, optionally a map of injected data and
+  returns the rendered template as a string."
   []
   (let [p (-> (core/get-env)
               (update-in [:dependencies] into deps)
@@ -27,24 +28,22 @@
                  ([tpl data inj]
                   (pod/with-call-in @p
                     (mrmcc3.boot-soy.impl/render ~tpl ~data ~inj))))
-        old-fileset (atom nil)]
+        prev-fileset (atom nil)]
     (core/with-pre-wrap fileset
-      (let [old @old-fileset
-            changed (->> (core/fileset-changed old fileset)
+      (let [prev    @prev-fileset
+            changed (->> (core/fileset-changed prev fileset)
                          core/input-files (core/by-ext [".soy"]))
-            added (->> (core/fileset-added old fileset)
-                       core/input-files (core/by-ext [".soy"]))
-            removed (->> (core/fileset-removed old fileset)
+            added   (->> (core/fileset-added prev fileset)
                          core/input-files (core/by-ext [".soy"]))
-            soy-src (->> fileset core/input-files (core/by-ext [".soy"])) ;; haha soy-src
-            paths (into #{} (map #(-> % core/tmp-file .getPath)) soy-src)]
-        (reset! old-fileset fileset)
+            removed (->> (core/fileset-removed prev fileset)
+                         core/input-files (core/by-ext [".soy"]))
+            paths   (->> fileset core/input-files (core/by-ext [".soy"])
+                         (map #(-> % core/tmp-file .getPath)))]
+        (reset! prev-fileset fileset)
+        (util/info "found %s soy files...\n" (count paths))
         (cond
           (or (seq added) (seq removed))
           (pod/with-call-in @p (mrmcc3.boot-soy.impl/set-files! ~paths))
           (seq changed)
           (pod/with-call-in @p (mrmcc3.boot-soy.impl/recompile-tofu!)))
-        (util/info "found %s soy files...\n" (count soy-src))
-        (-> (with-meta fileset {::render render})
-            (core/rm soy-src)
-            core/commit!)))))
+        (with-meta fileset {::render render})))))